requirements.txt
flask
pandas
scikit-learn
google-generativeai
markdown
werkzeug
python-dotenv
